# Context Menu: Code and performance updates for `useCopyPaste`