--- a/packages/story-editor/src/app/rightClickMenu/hooks/useCopyPasteActions.js
+++ b/packages/story-editor/src/app/rightClickMenu/hooks/useCopyPasteActions.js
@@ -32,6 +32,33 @@ import { getTextPresets } from '../../../utils/presetUtils';
 import { useHistory, useStory } from '../..';
 import { UNDO_HELP_TEXT } from './constants';
 
+function getTextElementStyles(element, animation) {
+  const { textStyles } = getTextPresets(
+    [element.styles],
+    {
+      textStyles: [],
+      colors: [],
+    },
+    PRESET_TYPES.STYLE
+  );
+  const { colors } = getTextPresets(
+    [element.styles],
+    {
+      textStyles: [],
+      colors: [],
+    },
+    PRESET_TYPES.COLOR
+  );
+  const { content, ...copiedElementStyles } = element.styles;
+  return {
+    ...copiedElementStyles,
+    ...textStyles[0],
+    ...colors[0].color,
+    animation,
+    border: copiedElementStyles.border || null,
+  };
+}
+
 /**
  * Creates the right click menu copy/paste actions.
  *
@@ -41,16 +68,20 @@ const useCopyPasteActions = () => {
   const undo = useHistory(({ actions }) => actions.undo);
   const {
     addAnimations,
-    copyElementById,
+    copySelectedElement,
     copiedElementState,
     selectedElement,
+    selectedElementType,
     selectedElementAnimations,
     updateSelectedElements,
+    isSelectedElementBackground,
   } = useStory(({ state, actions }) => ({
     addAnimations: actions.addAnimations,
     copiedElementState: state.copiedElementState,
-    copyElementById: actions.copyElementById,
+    copySelectedElement: actions.copySelectedElement,
     selectedElement: state.selectedElements?.[0],
+    selectedElementType: state.selectedElements?.[0].type,
+    isSelectedElementBackground: state.selectedElements?.[0].isBackground,
     selectedElementAnimations: state.selectedElementAnimations,
     updateSelectedElements: actions.updateSelectedElements,
   }));
@@ -65,11 +96,7 @@ const useCopyPasteActions = () => {
    * Copy the styles and animations of the selected element.
    */
   const handleCopyStyles = useCallback(() => {
-    if (!selectedElement?.id) {
-      return;
-    }
-
-    copyElementById({ elementId: selectedElement?.id });
+    copySelectedElement();
 
     showSnackbar({
       actionLabel: __('Undo', 'web-stories'),
@@ -80,8 +107,8 @@ const useCopyPasteActions = () => {
 
         trackEvent('context_menu_action', {
           name: 'undo_copy_styles',
-          element: selectedElement?.type,
-          isBackground: selectedElement?.isBackground,
+          element: selectedElementType,
+          isBackground: isSelectedElementBackground,
         });
       },
       actionHelpText: UNDO_HELP_TEXT,
@@ -89,10 +116,15 @@ const useCopyPasteActions = () => {
 
     trackEvent('context_menu_action', {
       name: 'copy_styles',
-      element: selectedElement?.type,
-      isBackground: selectedElement?.isBackground,
+      element: selectedElementType,
+      isBackground: isSelectedElementBackground,
     });
-  }, [copyElementById, selectedElement, showSnackbar]);
+  }, [
+    copySelectedElement,
+    showSnackbar,
+    selectedElementType,
+    isSelectedElementBackground,
+  ]);
 
   const pushUpdate = useCallback(
     (update, commitValues) => {
@@ -124,7 +156,7 @@ const useCopyPasteActions = () => {
   const handlePasteStyles = useCallback(() => {
     const id = selectedElement?.id;
 
-    if (!id || selectedElement?.type !== copiedElementState.type) {
+    if (!id || selectedElementType !== copiedElementState.type) {
       return;
     }
 
@@ -148,35 +180,14 @@ const useCopyPasteActions = () => {
       copiedElementState.type === 'text' &&
       copiedElementState.styles.content
     ) {
-      const { textStyles } = getTextPresets(
-        [copiedElementState.styles],
-        {
-          textStyles: [],
-          colors: [],
-        },
-        PRESET_TYPES.STYLE
+      const updatedElementStyles = getTextElementStyles(
+        copiedElementState,
+        oldAnimationToDelete
       );
-      const { colors } = getTextPresets(
-        [copiedElementState.styles],
-        {
-          textStyles: [],
-          colors: [],
-        },
-        PRESET_TYPES.COLOR
-      );
-      const { content, ...copiedElementStyles } = copiedElementState.styles;
-      const updatedElementStyles = {
-        ...copiedElementStyles,
-        ...textStyles[0],
-        ...colors[0].color,
-        animation: oldAnimationToDelete,
-        border: copiedElementStyles.border || null,
-      };
       handleApplyStyle(updatedElementStyles);
     } else {
       // Add styles and animations to element
       updateSelectedElements({
-        elementIds: [selectedElement.id],
         properties: (currentProperties) =>
           updateProperties(
             currentProperties,
@@ -200,8 +211,8 @@ const useCopyPasteActions = () => {
 
         trackEvent('context_menu_action', {
           name: 'undo_paste_styles',
-          element: selectedElement?.type,
-          isBackground: selectedElement?.isBackground,
+          element: selectedElementType,
+          isBackground: isSelectedElementBackground,
         });
       },
       actionHelpText: UNDO_HELP_TEXT,
@@ -209,8 +220,8 @@ const useCopyPasteActions = () => {
 
     trackEvent('context_menu_action', {
       name: 'paste_styles',
-      element: selectedElement?.type,
-      isBackground: selectedElement?.isBackground,
+      element: selectedElementType,
+      isBackground: isSelectedElementBackground,
     });
   }, [
     addAnimations,
@@ -220,6 +231,8 @@ const useCopyPasteActions = () => {
     selectedElementAnimations,
     showSnackbar,
     updateSelectedElements,
+    selectedElementType,
+    isSelectedElementBackground,
   ]);
 
   return {