--- a/packages/story-editor/src/app/story/useStoryReducer/reducers/index.js
+++ b/packages/story-editor/src/app/story/useStoryReducer/reducers/index.js
@@ -30,7 +30,7 @@ export { default as setBackgroundElement } from './setBackgroundElement';
 export { default as arrangeElement } from './arrangeElement';
 export { default as combineElements } from './combineElements';
 export { default as duplicateElementsById } from './duplicateElementsById';
-export { default as copyElementById } from './copyElementById';
+export { default as copySelectedElement } from './copySelectedElement';
 
 // Manipulate current page.
 export { default as setCurrentPage } from './setCurrentPage';