## Context
Part of looking through right click menu provider. Helps with a little code quality and some perf
<!-- What do we want to achieve with this PR? Why did we write this code? -->

## Summary
Some code quality improvements, and some perf improvements around primary actions:

**Selecting new Item**
| Before | After |
| ----- | ----- |
| ~9.2ms | ~0.1ms |
| <img width="494" alt="Screen Shot 2022-02-22 at 9 32 03 AM" src="https://user-images.githubusercontent.com/35983235/155165435-5edbf68e-31de-40ca-86f6-1b20501f3caa.png"> |  <img width="673" alt="Screen Shot 2022-02-22 at 9 30 42 AM" src="https://user-images.githubusercontent.com/35983235/155165465-07139fba-3554-4de4-9cf6-56d4413b914b.png"> |


**Drag to update element position**
| Before | After |
| ----- | ----- |
| ~2.9ms | ~0.1ms |
| <img width="624" alt="Screen Shot 2022-02-22 at 9 33 03 AM" src="https://user-images.githubusercontent.com/35983235/155165518-559a60c5-4b34-4b46-a428-a651733358b8.png"> |  <img width="670" alt="Screen Shot 2022-02-22 at 9 32 50 AM" src="https://user-images.githubusercontent.com/35983235/155165549-6a9f9eda-0b54-4401-8d16-3153ea9f77d8.png"> |


## Relevant Technical Choices
All perf improvements are around limiting some of the expensive editable text calls when they don't need to be recomputed. the editable text calls are called in some of the hooks of the right click menu provider needed to supply callbacks to some of the right click menu actions.

As far as code quality goes, I think I did all the tasks listed in the ticket except move a `pasteCopiedElementStyles` action into the reducer. At the end of the day, this wouldn't have improved anything re-render wise in the hook or amount of lines the callback is. After looking at it more in depth, it felt a little arbitrary since it wouldn't/couldn't handle pasting of text elements, so figured for now its fine to just leave all the pasting copied element styles logic contained in the same place in this callback defined in the `useCopyPasteActions` hook :shrug: . 
<!-- Please describe your changes. -->

## To-do

<!-- A list of things that need to be addressed in this PR or follow-up changes. -->

## User-facing changes
NA
<!--
Please describe your changes.
Include before/after screenshots or a short video.
-->

## Testing Instructions
NA
<!--
How can the changes in this PR be verified?
Please provide step-by-step instructions how to reproduce the issue, if applicable.
Write step-by-step test instructions aimed at non-tech-savvy users, even if the PR is not user-facing.
-->

<!-- ignore-task-list-start -->
- [x] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

1.


## Reviews

### Does this PR have a security-related impact?

<!-- Examples: new APIs, changes to KSES, etc.  -->

### Does this PR change what data or activity we track or use?

<!-- Examples: changes to telemetry, new third-party APIs -->

### Does this PR have a legal-related impact?

<!-- Examples: new images with unknown sources, new production dependencies with incompatible licenses -->

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [x] I have tested this code to the best of my abilities
- [x] I have verified accessibility to the best of my abilities ([docs](https://github.com/googleforcreators/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [x] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [x] This code is covered by automated tests (unit, integration, and/or e2e) to verify it works as intended ([docs](https://github.com/googleforcreators/web-stories-wp/tree/main/docs#testing))
- [x] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.

NOTE: One reference per line!

Example:

Fixes #123
Partially addresses #456
See #789
-->

Fixes #10459